requirements.txt
gradio
moviepy
pillow
pyttsx3
#Windows环境请安装pypiwin32
#pypiwin32
edge_tts
tqdm
qianfan
pydub
python-dotenv
librosa
soundfile
#安装LibreOffice
python-pptx
pdf2image
PyMuPDF
comtypes
Pillow
